requirements.txt
torch
torchvision
Pillow
transformers
requests 